Development: Split programming exercise service (#11028)
