+++ src/main/java/de/tum/cit/aet/artemis/programming/service/ProgrammingExerciseImportFromFileService.java
@@ -49,7 +49,9 @@ public class ProgrammingExerciseImportFromFileService {
 
     private static final Logger log = LoggerFactory.getLogger(ProgrammingExerciseImportFromFileService.class);
 
-    private final ProgrammingExerciseService programmingExerciseService;
+    private final ProgrammingExerciseCreationUpdateService programmingExerciseCreationUpdateService;
+
+    private final ProgrammingExerciseValidationService programmingExerciseValidationService;
 
     private final ZipFileService zipFileService;
 
@@ -67,10 +69,12 @@ public class ProgrammingExerciseImportFromFileService {
 
     private final BuildPlanRepository buildPlanRepository;
 
-    public ProgrammingExerciseImportFromFileService(ProgrammingExerciseService programmingExerciseService, ZipFileService zipFileService,
-            StaticCodeAnalysisService staticCodeAnalysisService, ProgrammingExerciseRepositoryService programmingExerciseRepositoryService, RepositoryService repositoryService,
-            GitService gitService, FileService fileService, ProfileService profileService, BuildPlanRepository buildPlanRepository) {
-        this.programmingExerciseService = programmingExerciseService;
+    public ProgrammingExerciseImportFromFileService(ProgrammingExerciseCreationUpdateService programmingExerciseCreationUpdateService,
+            ProgrammingExerciseValidationService programmingExerciseValidationService, ZipFileService zipFileService, StaticCodeAnalysisService staticCodeAnalysisService,
+            ProgrammingExerciseRepositoryService programmingExerciseRepositoryService, RepositoryService repositoryService, GitService gitService, FileService fileService,
+            ProfileService profileService, BuildPlanRepository buildPlanRepository) {
+        this.programmingExerciseCreationUpdateService = programmingExerciseCreationUpdateService;
+        this.programmingExerciseValidationService = programmingExerciseValidationService;
         this.zipFileService = zipFileService;
         this.staticCodeAnalysisService = staticCodeAnalysisService;
         this.programmingExerciseRepositoryService = programmingExerciseRepositoryService;
@@ -107,9 +111,9 @@ public class ProgrammingExerciseImportFromFileService {
             zipFileService.extractZipFileRecursively(exerciseFilePath);
             checkDetailsJsonExists(importExerciseDir);
             checkRepositoriesExist(importExerciseDir);
-            programmingExerciseService.validateNewProgrammingExerciseSettings(originalProgrammingExercise, course);
+            programmingExerciseValidationService.validateNewProgrammingExerciseSettings(originalProgrammingExercise, course);
             // TODO: creating the whole exercise (from template) is a bad solution in this case, we do not want the template content, instead we want the file content of the zip
-            newProgrammingExercise = programmingExerciseService.createProgrammingExercise(originalProgrammingExercise);
+            newProgrammingExercise = programmingExerciseCreationUpdateService.createProgrammingExercise(originalProgrammingExercise);
             if (Boolean.TRUE.equals(originalProgrammingExercise.isStaticCodeAnalysisEnabled())) {
                 staticCodeAnalysisService.createDefaultCategories(newProgrammingExercise);
             }
