+++ src/test/java/de/tum/cit/aet/artemis/programming/AbstractProgrammingIntegrationJenkinsLocalVCTest.java
@@ -27,6 +27,7 @@ import de.tum.cit.aet.artemis.programming.repository.BuildPlanRepository;
 import de.tum.cit.aet.artemis.programming.repository.ProgrammingExerciseBuildConfigRepository;
 import de.tum.cit.aet.artemis.programming.service.ConsistencyCheckTestService;
 import de.tum.cit.aet.artemis.programming.service.ProgrammingExerciseService;
+import de.tum.cit.aet.artemis.programming.service.ProgrammingExerciseValidationService;
 import de.tum.cit.aet.artemis.programming.service.ProgrammingLanguageFeatureService;
 import de.tum.cit.aet.artemis.programming.service.jenkins.JenkinsAuthorizationInterceptor;
 import de.tum.cit.aet.artemis.programming.service.jenkins.JenkinsInternalUrlService;
@@ -119,6 +120,9 @@ public abstract class AbstractProgrammingIntegrationJenkinsLocalVCTest extends A
 
     @Autowired
     protected ProgrammingExerciseService programmingExerciseService;
+
+    @Autowired
+    protected ProgrammingExerciseValidationService programmingExerciseValidationService;
 
     @Autowired
     protected ProgrammingLanguageFeatureService programmingLanguageFeatureService;
